refactor: Moved footer outside of container components
