+++ src/index.js
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux';
 import { ConnectedRouter } from 'react-router-redux';
 
 import store, { history } from './store';
-import App from './App';
+import App from './components/App';
 import './index.css';
 
 const target = document.getElementById('root');
